+++ json_database/exceptions.py
@@ -4,3 +4,7 @@ class InvalidItemID(ValueError):
 
 class DatabaseNotCommitted(FileNotFoundError):
     """ Database has not been saved in yet """
+
+
+class SessionError(RuntimeError):
+    """ Could not commit database"""