+++ CHANGELOG.md
@@ -15,5 +15,5 @@ current dev branch
 initial release
 
 
-[unreleased]: https://github.com/OpenJarbas/jarbas_utils/tree/dev
+[unreleased]: https://github.com/OpenJarbas/json_database/tree/dev
 [0.1.0]: https://github.com/OpenJarbas/json_database/tree/0.1.0